Assistant checkpoint: Fix DOCTYPE and login redirection
Assistant generated file changes:
- application/view/_templates/header.php: Add proper DOCTYPE and fix HTML structure
- application/controller/LoginController.php: Fix login redirection

---

User prompt:

please look into the login files or into the header Idk but there is a problem with redirecting and could be that the problem is not even made in redirecting files
afte login I don´t land on the dashboard whatever I try fix that!!

and this too

One or more documents in this page is in Quirks Mode, which will render the affected document(s) with quirks incompatible with the current HTML and CSS specifications.

Quirks Mode exists mostly due to historical reasons. If this is not intentional, you can add or modify the DOCTYPE to be `<!DOCTYPE html>` to render the page in No Quirks Mode.